Update AddInfo code with AddInfo2 code
